requirements.txt
gspread
google-auth
openai
python-dotenv
python-telegram-bot==21.7
